+++ packages/editor/editor-bitbucket-transformer/src/serializer.ts
@@ -274,14 +274,17 @@ const editorNodes = {
       state.render(child, node, i);
     }
   },
-  mediaSingle(state: MarkdownSerializerState, node: PMNode) {
+  mediaSingle(state: MarkdownSerializerState, node: PMNode, parent: PMNode) {
     for (let i = 0; i < node.childCount; i++) {
       const child = node.child(i);
       state.render(child, node, i);
+      if (!parent.type.name.startsWith('table')) {
+        state.write('\n');
+      }
     }
   },
   media(state: MarkdownSerializerState, node: PMNode) {
-    state.write('![](' + node.attrs.url + ')\n');
+    state.write('![](' + node.attrs.url + ')');
   },
   image(state: MarkdownSerializerState, node: PMNode) {
     // Note: the 'title' is not escaped in this flavor of markdown.