+++ packages/media/media-viewer/__tests__/newgen/viewers/doc/pdfRenderer.spec.tsx
@@ -2,18 +2,18 @@ import * as React from 'react';
 import { mount } from 'enzyme';
 import Button from '@atlaskit/button';
 import {
-  PDFViewer,
+  PDFRenderer,
   pdfViewerClassName,
-} from '../../../../src/newgen/viewers/pdf/pdfComponent';
+} from '../../../../src/newgen/viewers/doc/pdfRenderer';
 import { ZoomControls } from '../../../../src/newgen/zoomControls';
 
 function createFixture() {
   const onClose = jest.fn();
-  const el = mount(<PDFViewer doc={null} onClose={onClose} />);
+  const el = mount(<PDFRenderer src={''} onClose={onClose} />);
   return { el, onClose };
 }
 
-describe('PDFViewer', () => {
+describe('PDFRenderer', () => {
   it('supports zooming', async () => {
     const { el } = createFixture();
 
@@ -27,7 +27,7 @@ describe('PDFViewer', () => {
     expect(el.state('zoom')).toBeLessThan(100);
   });
 
-  it('MSW-700: clicking on background of PDFViewer does not close it', () => {
+  it('MSW-700: clicking on background of DocViewer does not close it', () => {
     const { el, onClose } = createFixture();
     el.find(`.${pdfViewerClassName}`).simulate('click');